apps/worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worker"]

[project]
name = "atlas-worker"
version = "0.0.1"
description = "Atlas Cloud Cost Optimization Celery Worker"
requires-python = ">=3.11"
dependencies = [
    # Task queue
    "celery[redis]>=5.4.0",
    "kombu>=5.3.4",

    # Database (shared with API)
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",

    # Cloud SDKs
    "boto3>=1.42.79,<1.42.85",

    # Redis (standalone client — separate from celery[redis] broker transport)
    "redis>=5.0.0",

    # Forecasting
    "statsmodels>=0.14.2",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",

    # Utilities
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.3",
    "httpx>=0.27.0",
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",        # parallel test execution
    "pytest-timeout>=2.3.1",      # per-test time limits
    "pytest-mock>=3.14.0",
    "celery[pytest]>=5.4.0",
    "fakeredis>=2.23.0",          # in-process Redis
    # Load testing
    "locust>=2.24.0",
    "ruff>=0.6.0",
]

# ─── Ruff ─────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["worker"]

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501", "F841"]

[tool.ruff.lint.isort]
known-first-party = ["worker"]
force-single-line = true

# ─── Black ────────────────────────────────────────────────────────────────────
[tool.black]
target-version = ["py311"]
line-length = 100

# ─── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

# ─── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode    = "auto"
testpaths       = ["tests"]
log_cli         = true
log_cli_level   = "WARNING"
filterwarnings  = [
    "ignore::DeprecationWarning:statsmodels.*",
    "ignore::FutureWarning:pandas.*",
    "ignore::UserWarning:statsmodels.*",
]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-q",
]
markers = [
    "unit: fast, no I/O — runs in CI without any external services",
    "integration: tests multiple real components together; mocks I/O boundaries",
    "load: Locust / throughput benchmarks — not run in CI by default",
    "slow: tests that take > 5 s — skipped unless --run-slow passed",
    "determinism: verifies identical output on repeated runs",
]

# ─── Coverage ─────────────────────────────────────────────────────────────────
[tool.coverage.run]
source          = ["worker"]
branch          = false
omit            = [
    "*/tests/*",
    "worker/__init__.py",
    "worker/config.py",
    "worker/db.py",
    "worker/dlq.py",
    "worker/forecasting/store.py",
    "worker/main.py",
    "worker/optimization/__init__.py",
    "worker/optimization/benchmark.py",
    "worker/optimization/engine.py",
    "worker/optimization/scorer.py",
    "worker/optimization/solver.py",
    "worker/rate_limiter.py",
    "worker/redis_client.py",
    "worker/telemetry.py",
    "worker/tasks/cost_sync.py",   # scaffolding, not yet implemented
    "worker/tasks/dlq_processor.py",
    "worker/tasks/forecast.py",
    "worker/tasks/optimize.py",
    "worker/tasks/recommendations.py",
]

[tool.coverage.report]
fail_under      = 80
show_missing    = true
skip_covered    = false
exclude_lines   = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "def __repr__",
]

[tool.coverage.html]
directory = "htmlcov"
